refactor: extract DefaultRoles
(cherry picked from commit 2063b1ab6571f995c6c208d269efca9f9d285951)
